renamed to azurecs
