+++ modules/azureCS/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "AzureCS",
+  "version": "1.0.0",
+  "description": "Microsoft Azure Cognitive Services Custom Module (Language, Search API)",
+  "main": "build/module.js",
+  "scripts": {
+    "build": "tsc -p ."
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Cognigy/CustomModules"
+  },
+  "keywords": [
+    "Microsoft",
+    "Azure",
+    "Cognitive",
+    "Services"
+  ],
+  "author": "Cognigy GmbH",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/node": "^10.12.20",
+    "tslint": "^5.12.1",
+    "typescript": "^3.3.3"
+  },
+  "dependencies": {
+    "https": "1.0.0",
+    "request": "2.88.0",
+    "uuid": "3.3.2"
+  }
+}
