+++ modules/azureCS/tsconfig.json
@@ -1,0 +1,24 @@
+﻿{
+  "compileOnSave": true,
+  "compilerOptions": {
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "target": "es2017", 
+    "module": "commonjs",
+    "rootDir": "src", 
+    "outDir": "build",
+    "sourceMap": true,
+    "moduleResolution": "node",
+    "watch": false
+  },
+  "exclude": [
+    "node_modules",
+    "old",
+    "build",
+    "./vscode"
+  ],
+  "include": [
+    "../../common/cognigy.d.ts",
+    "src/**/*"
+  ]
+}
