+++ modules/azureCS/tslint.json
@@ -1,0 +1,33 @@
+{
+    "rules": {
+        "prefer-const": true,
+        "prefer-for-of": true,
+        "only-arrow-functions": [true, "allow-declarations", "allow-named-functions"],
+        "no-var-keyword": true,
+        "no-var-requires": false,
+        "comment-format": [true, "check-space"],
+        "no-duplicate-variable": true,
+        "no-eval": false,
+        "no-internal-module": true,
+        "no-trailing-whitespace": true,
+        "one-line": [true, "check-open-brace", "check-whitespace", "check-catch", "check-finally", "check-else"],
+        "semicolon": [true, "always"],
+        "triple-equals": [true, "allow-null-check"],
+        "typedef-whitespace": [true, {
+            "call-signature": "nospace",
+            "index-signature": "nospace",
+            "parameter": "nospace",
+            "property-declaration": "nospace",
+            "variable-declaration": "nospace"
+        }],
+        "typedef": [true, "call-signature", "parameter", "member-variable-declaration"],
+        "variable-name": [true, "ban-keywords"],
+        "whitespace": [true,
+            "check-branch",
+            "check-decl",
+            "check-operator",
+            "check-separator",
+            "check-type"
+        ]
+    }
+}
